+++ artifact_gen/lib/component/attach.dart
@@ -21,6 +21,7 @@ class $ArtifactAttachComponent implements $ArtifactBuilderOutput {
     ConstructorElement ctor,
     List<FormalParameterElement> params,
     BuildStep step,
+    List<String>? eFields,
   ) async {
     StringBuffer buf = StringBuffer();
     List<Uri> importUris = <Uri>[];