+++ artifact_gen/lib/component/inatance.dart
@@ -14,6 +14,7 @@ class $ArtifactInstanceComponent implements $ArtifactBuilderOutput {
     ConstructorElement ctor,
     List<FormalParameterElement> params,
     BuildStep step,
+    List<String>? eFields,
   ) async {
     StringBuffer buf = StringBuffer();
     List<Uri> importUris = <Uri>[];